+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.7.0 (in development)
+-----------------------
+- `properties2json` and `json2properties` now preserve the input ordering of
+  keys by default; use the new `--sort-keys` option to get the old behavior
+
 v0.6.1 (2020-03-09)
 -------------------
 - Require `javaproperties` 0.7.\*