+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+v0.7.0 (in development)
+-----------------------
+- :program:`properties2json` and :program:`json2properties` now preserve the
+  input ordering of keys by default; use the new ``--sort-keys`` option to get
+  the old behavior
+
 v0.6.1 (2020-03-09)
 -------------------
 - Require ``javaproperties`` 0.7.\*